+++ apps/api/src/app/app.module.ts
@@ -34,6 +34,7 @@ import { RedisCacheModule } from './redis-cache/redis-cache.module';
 import { SubscriptionModule } from './subscription/subscription.module';
 import { SymbolModule } from './symbol/symbol.module';
 import { UserModule } from './user/user.module';
+import { PlatformModule } from './platform/platform.module';
 
 @Module({
   imports: [
@@ -63,6 +64,7 @@ import { UserModule } from './user/user.module';
     InfoModule,
     LogoModule,
     OrderModule,
+    PlatformModule,
     PortfolioModule,
     PrismaModule,
     RedisCacheModule,