+++ apps/api/src/app/platform/create-platform.dto.ts
@@ -1,0 +1,9 @@
+import { IsString } from 'class-validator';
+
+export class CreatePlatformDto {
+  @IsString()
+  name: string;
+
+  @IsString()
+  url: string;
+}