+++ apps/api/src/app/platform/update-platform.dto.ts
@@ -1,0 +1,12 @@
+import { IsString } from 'class-validator';
+
+export class UpdatePlatformDto {
+  @IsString()
+  id: string;
+
+  @IsString()
+  name: string;
+
+  @IsString()
+  url: string;
+}